Complete splitting up the mega trait.
There are now many more manageable traits instead of one monolithic
trait. There are still a few bits to shuffle around and tests to align
with the code, but that can be done separately I think.
