+++ src/RelativeKeywordTrait.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
+ *
+ * Licensed under The MIT License
+ * Redistributions of files must retain the above copyright notice. Provides various operator methods for datetime
+ * objects.
+ *
+ * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
+ * @copyright     Copyright (c) Brian Nesbitt <brian@nesbot.com>
+ * @link          http://cakephp.org CakePHP(tm) Project
+ * @license       http://www.opensource.org/licenses/mit-license.php MIT License
+ */
+namespace Cake\Chronos;
+
+/**
+ * Provides methods for testing if strings contain relative keywords.
+ */
+trait RelativeKeywordTrait
+{
+    /**
+     * Terms used to detect if a time passed is a relative date for testing purposes
+     *
+     * @var array
+     */
+    protected static $relativeKeywords = [
+        'this',
+        'next',
+        'last',
+        'tomorrow',
+        'yesterday',
+        '+',
+        '-',
+        'first',
+        'last',
+        'ago',
+    ];
+
+    /**
+     * Determine if there is a relative keyword in the time string, this is to
+     * create dates relative to now for test instances. e.g.: next tuesday
+     *
+     * @param string $time The time string to check.
+     * @return bool true if there is a keyword, otherwise false
+     */
+    public static function hasRelativeKeywords($time)
+    {
+        // skip common format with a '-' in it
+        if (preg_match('/[0-9]{4}-[0-9]{1,2}-[0-9]{1,2}/', $time) !== 1) {
+            foreach (static::$relativeKeywords as $keyword) {
+                if (stripos($time, $keyword) !== false) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+}
